requirements.txt
Flask==3.1.1
openai==1.81.0
python-dotenv==1.0.1